Revert "Started p4. Refactored a lot of p3 into object.h/c so that I can use it in p4."
This reverts commit f56491a4c6ea6b308b3e03b4928d41dd5806137d.

Reverting to get rid of all the output files.

+++ hw2.h
@@ -2,17 +2,13 @@
  * hw2.h
  * Contains useful typedefs and macros.
  */
-
 #ifndef HW2_H_
 #define HW2_H_
 
 #define BINARY 1
 #define DIM 2
 #define PIXEL_MAX 255
-#define DEG_PER_RAD (180.0f/PI)
-
-#include <stdlib.h>
-#include <stdio.h>
+#define DEG_PER_RAD 180.0/PI
 
 #include "vision_utilities.h"
 
@@ -28,7 +24,7 @@ typedef struct LabelMap {
 /** Holds an Object's second moment */
 typedef struct SecondMoment {
 	int a, b, c;
-	float thetaMin, thetaMax;
+	double thetaMin, thetaMax;
 } SecondMoment;
 
 /** Holds data on Objects found in an image */
